+++ src/service/transaction/placeOrderInProgress/potentialActions/givePointAward.ts
@@ -15,6 +15,7 @@ export async function createGivePointAwardActions(params: {
         givePointAwardParams.forEach((givePointAwardParam) => {
             const amount = givePointAwardParam.object?.amount;
 
+            const toLocationType = givePointAwardParam.object?.toLocation?.typeOf;
             const accountNumber = givePointAwardParam.object?.toLocation?.accountNumber;
             const accountType = givePointAwardParam.object?.toLocation?.accountType;
             const description = givePointAwardParam.object?.description;
@@ -31,7 +32,7 @@ export async function createGivePointAwardActions(params: {
                         typeOf: factory.action.transfer.give.pointAward.ObjectType.PointAward,
                         amount: amount,
                         toLocation: {
-                            typeOf: factory.chevre.paymentMethodType.Account,
+                            typeOf: toLocationType,
                             accountNumber: accountNumber,
                             accountType: accountType
                         },