+++ CHANGELOG.md
@@ -10,6 +10,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ### Changed
 
+- factory.chevre.paymentMethodType.Accountへの依存を排除
+
 ### Deprecated
 
 ### Removed